submodules/NASA-IMU-Wrapper/requirements.txt
Adafruit-Blinka==8.20.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-lis3mdl==1.1.24
adafruit-circuitpython-lsm6ds==4.5.10
adafruit-circuitpython-register==1.9.16
adafruit-circuitpython-requests==2.0.0
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.47.0
Adafruit-PureIO==1.1.11
AHRS==0.3.1
board==1.0
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.0
idna==3.4
importlib-resources==6.0.0
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
pyftdi==0.54.0
pyparsing==3.0.9
PyQt6-sip==13.5.1
pyqtgraph==0.13.3
pyserial==3.5
python-dateutil==2.8.2
pyusb==1.2.1
remote-plot==1.2.1
requests==2.31.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.16.0
squaternion==0.3.4
sysv-ipc==1.1.0
typing_extensions==4.7.1
urllib3==2.0.3
zipp==3.16.2